requirements.txt
feedparser==6.0.10
langchain==0.0.129
openai==0.27.2
pandas==1.5.3
pinecone-client==2.2.1
python-box==7.0.1
python-dotenv==1.0.0
streamlit==1.20.0
streamlit-extras==0.2.6